adding third party jars

+++ src/main/java/com/demandcube/demandspike/kafka/producer/Statistic.java
@@ -72,7 +72,7 @@ public class Statistic implements Runnable {
 				}
 				eventsSentPerPeriod = eventsSentTotal - previousEventsSentTotal;
 				bytesSentPerPeriod = bytesSentTotal - previousBytesSentTotal;
-				eventsSentAvg = eventsSentTotal / secondCount;
+				eventsSentAvg = (double) eventsSentTotal / secondCount;
 
 				NumberFormat formatter = new DecimalFormat("#0.00");
